feat: BPMをfloat対応にして推薦の除外しきい値を>6に厳格化
- calc_bpm_score をint/float両対応にし、差>6.0でNoneを返すよう変更

- Track.bpmをFloat型に変更（冪等マイグレーション 002_bpm_float.py 追加）

- recommend / set_generator / API の呼び出し側でNone候補をフィルタ

- フォーム入力にstep=0.1を追加、CSVインポートもfloatを保持

- 境界値とfloat入力のテストを追加（全132件PASS）

+++ app/models/track.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, CheckConstraint, ForeignKey
+from sqlalchemy import Column, Integer, Float, String, CheckConstraint, ForeignKey
 from sqlalchemy.orm import relationship
 from app.extensions import Base
 
@@ -14,8 +14,9 @@ class Track(Base):
 
     artist = Column(String(255), nullable=False)
 
+    # BPM は小数も許容（例: 80.5）。整数値もそのまま保存可能。
     bpm = Column(
-        Integer,
+        Float,
         nullable=False
     )
